+++ playbooks/installer/ospd/overcloud/pre.yml
@@ -11,8 +11,8 @@
             msg: "Stopping before building/importing the images per user request"
         when: break is defined and break == "before_images"
   roles:
-      - {role: ospd/overcloud/images/build, when: installer.images.task == "build"}
-      - {role: ospd/overcloud/images/import, when: installer.images.task == "import"}
+      - {role: ospd/overcloud/images/build, when: installer.overcloud.images.task == "build"}
+      - {role: ospd/overcloud/images/import, when: installer.overcloud.images.task == "import"}
   tasks:
       - name: upload the overcloud images to glance
         shell: "source ~/stackrc; openstack overcloud image upload"
@@ -63,7 +63,7 @@
 
       - name: Print the MAC addresses for instackenv.json
         debug:
-            msg: "MAC address for instackenv.json: {{ hostvars[item]['ansible_%s' % installer.undercloud.config.local_interface].macaddress }}"
+            msg: "MAC address for instackenv.json: {{ hostvars[item]['ansible_%s' % installer.undercloud.network.local_interface].macaddress }}"
         with_items: groups['openstack_nodes'] | difference(["virthost"]) | difference(["undercloud"])
         ignore_errors: true
 
@@ -155,7 +155,7 @@
         shell: "source ~/stackrc; openstack baremetal introspection bulk start"
 
       - name: get subnet ID to update neutron's DNS server
-        shell: "source ~/stackrc; neutron subnet-list | grep {{ installer.undercloud.config.network_cidr }} | awk '{print $2;}'"
+        shell: "source ~/stackrc; neutron subnet-list | grep {{ installer.undercloud.network.network_cidr }} | awk '{print $2;}'"
         register: subnet_id
 
       - name: get the nameserver
