+++ roles/ospd/overcloud/network/isolation/single-nic-vlans/tasks/main.yml
@@ -4,12 +4,12 @@
       dest: "{{ installer.overcloud.template_base }}/isolation_params.yml"
 
 - name: append our params to the network isolation file
-  shell: "cat {{ installer.overcloud.template_base }}/isolation_params.yml >> {{ installer.overcloud.template_base }}/{{ installer.overcloud.network.template.file }}"
+  shell: "cat {{ installer.overcloud.template_base }}/isolation_params.yml >> {{ installer.overcloud.template_base }}/{{ installer.overcloud.network.isolation.file }}"
 
 - name: append the network isolation template line to the base overcloud deploy script
   lineinfile:
       dest: "~/overcloud_deploy.sh"
-      line: '                       -e {{ installer.overcloud.template_base }}/{{ installer.overcloud.network.template.file }} \\'
+      line: '                       -e {{ installer.overcloud.template_base }}/{{ installer.overcloud.network.isolation.file }} \\'
 
 - name: append the net-single-nic-vlans template line base overcloud deploy script
   lineinfile:
@@ -17,11 +17,11 @@
       line: '                       -e {{ installer.overcloud.template_base }}/{{ installer.overcloud.network.isolation.file }} \\'
 
 - name: add port to br-ctlplane for ipv6 over vlan support
-  shell: "sudo ovs-vsctl add-port br-ctlplane vlan{{ installer.overcloud.network.template.content.parameter_defaults.ExternalNetworkVlanID }} tag={{ installer.overcloud.network.template.content.parameter_defaults.ExternalNetworkVlanID }} -- set interface vlan{{ installer.overcloud.network.template.content.parameter_defaults.ExternalNetworkVlanID }} type=internal"
+  shell: "sudo ovs-vsctl add-port br-ctlplane vlan{{ installer.overcloud.network.isolation.template.content.parameter_defaults.ExternalNetworkVlanID }} tag={{ installer.overcloud.network.isolation.template.content.parameter_defaults.ExternalNetworkVlanID }} -- set interface vlan{{ installer.overcloud.network.isolation.template.content.parameter_defaults.ExternalNetworkVlanID }} type=internal"
   when: installer.overcloud.network.backend == 'vlan'
 
 - name: add the ipv6 address
-  shell: "sudo ip l set dev vlan{{ installer.overcloud.network.template.content.parameter_defaults.ExternalNetworkVlanID }} up; sudo ip addr add {{ installer.overcloud.network.template.content.parameter_defaults.ExternalInterfaceDefaultRoute }}/64 dev vlan{{ installer.overcloud.network.template.content.parameter_defaults.ExternalNetworkVlanID }}"
+  shell: "sudo ip l set dev vlan{{ installer.overcloud.network.isolation.template.content.parameter_defaults.ExternalNetworkVlanID }} up; sudo ip addr add {{ installer.overcloud.network.isolation.template.content.parameter_defaults.ExternalInterfaceDefaultRoute }}/64 dev vlan{{ installer.overcloud.network.isolation.template.content.parameter_defaults.ExternalNetworkVlanID }}"
   register: set_ipv6
   failed_when: "set_ipv6.stderr != '' and 'RTNETLINK answers: File exists' not in set_ipv6.stderr"
   when: installer.overcloud.network.backend == 'vlan' and installer.overcloud.network.protocol == 'ipv6'
