+++ roles/ospd/overcloud/network/isolation/single-nic-vlans/templates/isolation.yml.j2
@@ -1,1 +1,1 @@
-{{ installer.overcloud.network.template.content | to_yaml }}
+{{ installer.overcloud.network.isolation.template.content | to_yaml }}
