+++ roles/ospd/overcloud/network/isolation/three-nics-vlans/templates/network-environment.yaml.j2
@@ -1,1 +1,1 @@
-{{ installer.overcloud.network.template.content | to_yaml }}
+{{ installer.overcloud.network.isolation.template.content | to_yaml }}
