+++ settings/base.spec
