+++ settings/installer/ospd/firewall/rules/default.yml
@@ -1,0 +1,11 @@
+---
+- ' -A INPUT -m state --state ESTABLISHED,RELATED -m comment --comment "001 QA
+  incoming related session exist the host interface MGMT" -j ACCEPT'
+- ' -A INPUT -p tcp -m tcp --dport 22 -m comment --comment "002 QA incoming SSH"
+  -j ACCEPT'
+- ' -A INPUT -p icmp -m comment --comment "003 QA incoming ICMP" -j ACCEPT'
+- ' -A INPUT -i lo -m comment --comment "004 QA incoming loopback" -j ACCEPT'
+- ' -A INPUT -p udp --sport 53 -m comment --comment "005 QA incoming DNS" -j ACCEPT'
+- ' -A INPUT -p udp --sport 123 -m comment --comment "006 QA incoming NTP" -j
+  ACCEPT'
+- ' -P INPUT DROP'
