--- settings/installer/ospd/images/import.yml
@@ -1,9 +1,0 @@
----
-
-installer:
-    images:
-        base_url: "{{ !lookup private.installer.images.{{ !lookup product.build }} }}"
-        task: import
-
-defaults:
-    version: 7
