--- settings/installer/ospd/images/import/version/7.yml
@@ -1,9 +1,0 @@
----
-
-installer:
-    images:
-        overcloud:
-            files:
-                discovery: "discovery-ramdisk.tar"
-                deployment: "deploy-ramdisk-ironic.tar"
-                overcloud: "overcloud-full.tar"
