--- settings/installer/ospd/images/import/version/8.yml
@@ -1,8 +1,0 @@
----
-
-installer:
-    images:
-        overcloud:
-            files:
-                discovery: "ironic-python-agent.tar"
-                overcloud: "overcloud-full.tar"
