--- settings/installer/ospd/network/isolation.yml
@@ -1,12 +1,0 @@
----
-#TODO: move everything out of isolation, have isolation as it's own option
-installer:
-    overcloud:
-        network:
-            isolation:
-                enable: "yes"
-
-defaults:
-    protocol: ipv4
-    type: "three-nics-vlan"
-    variant: vxlan
