--- settings/installer/ospd/network/isolation/protocol/ipv4.yml
@@ -1,30 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            protocol: ipv4
-            template:
-                content:
-                    parameter_defaults:
-                        InternalApiNetCidr: 172.17.1.0/24
-                        InternalApiAllocationPools: [{'start': '172.17.1.10', 'end': '172.17.1.200'}]
-                        InternalApiNetworkVlanID: 101
-                        TenantNetCidr: 172.17.2.0/24
-                        TenantAllocationPools: [{'start': '172.17.2.10', 'end': '172.17.2.200'}]
-                        TenantNetworkVlanID: 201
-                        StorageNetCidr: 172.17.3.0/24
-                        StorageAllocationPools: [{'start': '172.17.3.10', 'end': '172.17.3.200'}]
-                        StorageNetworkVlanID: 301
-                        StorageMgmtNetCidr: 172.17.4.0/24
-                        StorageMgmtAllocationPools: [{'start': '172.17.4.10', 'end': '172.17.4.200'}]
-                        StorageMgmtNetworkVlanID: 401
-                        ExternalNetCidr: 192.168.1.0/24
-                        ExternalAllocationPools: [{'start': '192.168.1.101', 'end': '192.168.1.149'}]
-                        ExternalInterfaceDefaultRoute: 192.168.1.1
-                        ExternalNetworkVlanID: 501
-                        ControlPlaneSubnetCidr: "24"
-                        ControlPlaneDefaultRoute: 172.16.0.1
-                        EC2MetadataIp: 172.16.0.1
-                        DnsServers: ['192.168.1.1', '8.8.8.8']
-                        NeutronExternalNetworkBridge: "''"
