--- settings/installer/ospd/network/isolation/protocol/ipv6.yml
@@ -1,31 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            backend: vlan
-            protocol: ipv6
-            template:
-                content:
-                    parameter_defaults:
-                        InternalApiNetCidr: 'fd00:fd00:fd00:2000::/64'
-                        InternalApiAllocationPools: [{'start': 'fd00:fd00:fd00:2000::10', 'end': 'fd00:fd00:fd00:2000:ffff:ffff:ffff:fffe'}]
-                        InternalApiNetworkVlanID: 101
-                        TenantNetCidr: 172.17.2.0/24
-                        TenantAllocationPools: [{'start': '172.17.2.10', 'end': '172.17.2.200'}]
-                        TenantNetworkVlanID: 201
-                        StorageNetCidr: 'fd00:fd00:fd00:3000::/64'
-                        StorageAllocationPools: [{'start': 'fd00:fd00:fd00:3000::10', 'end': 'fd00:fd00:fd00:3000:ffff:ffff:ffff:fffe'}]
-                        StorageNetworkVlanID: 301
-                        StorageMgmtNetCidr: 'fd00:fd00:fd00:4000::/64'
-                        StorageMgmtAllocationPools: [{'start': 'fd00:fd00:fd00:4000::10', 'end': 'fd00:fd00:fd00:4000:ffff:ffff:ffff:fffe'}]
-                        StorageMgmtNetworkVlanID: 401
-                        ExternalNetCidr: '2620:52:0:13b8::/64'
-                        ExternalAllocationPools: [{'start': '2620:52:0:13b8:5054:ff:fe3e:1', 'end': '2620:52:0:13b8:5054:ff:fe3e:9'}]
-                        ExternalInterfaceDefaultRoute: 2620:52:0:13b8::fe
-                        ExternalNetworkVlanID: 501
-                        ControlPlaneSubnetCidr: "24"
-                        ControlPlaneDefaultRoute: 172.16.0.1
-                        EC2MetadataIp: 172.16.0.1
-                        DnsServers: ['192.168.1.1', '8.8.8.8']
-                        NeutronExternalNetworkBridge: "''"
