--- settings/installer/ospd/network/isolation/type/single-nic-vlans.yml
@@ -1,9 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            isolation:
-                enable: "yes"
-                type: "single-nic-vlans"
-                file: "environments/net-single-nic-with-vlans.yaml"
