--- settings/installer/ospd/network/isolation/type/three-nics-vlans.yml
@@ -1,9 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            isolation:
-                enable: "yes"
-                type: "three-nics-vlans"
-                file: "environments/net-single-nic-with-vlans.yaml"
