--- settings/installer/ospd/network/isolation/variant/gre.yml
@@ -1,6 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            backend: gre
