--- settings/installer/ospd/network/isolation/variant/sriov.yml
@@ -1,7 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            backend: sriov
-
