--- settings/installer/ospd/network/isolation/variant/vlan.yml
@@ -1,6 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            backend: vlan
