--- settings/installer/ospd/network/isolation/variant/vxlan.yml
@@ -1,6 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            backend: vxlan
