--- settings/installer/ospd/network/no-isolation.yml
@@ -1,8 +1,0 @@
----
-
-installer:
-    overcloud:
-        network:
-            isolation:
-                enable: "no"
-            backend: vxlan
