+++ settings/installer/ospd/ospd.yml
@@ -1,0 +1,43 @@
+---
+
+installer:
+    type: ospd
+    short_type: ospd
+
+    overcloud:
+      template_base: ~/my_templates
+
+job:
+  archive:
+  - /var/log/
+  - /etc/yum.repos.d
+  - /etc/selinux
+  - /root/
+  - /var/lib/mysql
+  - /etc/nova
+  - /etc/ceilometer
+  - /etc/cinder
+  - /etc/glance
+  - /etc/keystone
+  - /etc/neutron
+  - /etc/swift
+  - /etc/heat
+  - /etc/libvirt
+  - /etc/rabbitmq
+  - /etc/ntp
+  - /etc/puppet
+  - /etc/qpid
+  - /etc/qpidd.conf
+  - /home/stack/.instack/install-undercloud.log
+  - /home/stack/stackrc
+  - /home/stack/overcloudrc
+  - /home/stack/*.log
+  - /home/stack/*.json
+  - /home/stack/*.conf
+  - /home/stack/*.yml
+  - /home/stack/deploy-overcloudrc
+  - /home/stack/network-environment.yaml
+  - /home/stack/tempest/*.xml
+  - /home/stack/tempest/*.log
+  - /home/stack/tempest/etc/*.conf
+  - /usr/share/openstack-tripleo-heat-templates
