+++ settings/installer/ospd/overcloud/images/files/default_import.yml
@@ -1,0 +1,3 @@
+discovery: "discovery-ramdisk.tar"
+deployment: "deploy-ramdisk-ironic.tar"
+overcloud: "overcloud-full.tar"
