+++ settings/installer/ospd/overcloud/network/isolation/single-nic-vlans.yml
@@ -1,0 +1,4 @@
+---
+enable: "yes"
+type: "single-nic-vlans"
+file: "environments/net-single-nic-with-vlans.yaml"
