+++ settings/installer/ospd/overcloud/network/isolation/three-nics-vlans.yml
@@ -1,0 +1,4 @@
+---
+enable: "yes"
+type: "three-nics-vlans"
+file: "environments/net-single-nic-with-vlans.yaml"
