+++ settings/installer/ospd/overcloud/storage/ceph.yml
@@ -1,0 +1,8 @@
+---
+external: "no"
+template:
+    content:
+        parameter_defaults:
+            CinderRbdPoolName: "volumes"
+            NovaRbdPoolName: "vms"
+            GlanceRbdPoolName: "images"
