+++ settings/installer/ospd/product/rhos-latest.yml
@@ -1,0 +1,24 @@
+---
+rpmrepo:
+  RedHat: http://rhos-release.virt.bos.redhat.com/repos/rhos-release/
+name: rhos
+short_name: rh
+config:
+  enable_epel: n
+full_version: '7.0'
+repo:
+  location: bos
+  mirror: download.eng.bos.redhat.com
+  release: latest
+  rhos_release:
+    extra_args:
+    - -p
+    - latest
+  type: puddle
+  short_type: pud
+version:
+  major: 7
+  minor: 0
+  code_name: kilo
+build: latest
+rpm: http://rhos-release.virt.bos.redhat.com/repos/rhos-release/rhos-release-latest.noarch.rpm
