--- settings/installer/ospd/ssl/no.yml
@@ -1,5 +1,0 @@
----
-
-installer:
-    overcloud:
-        use_ssl: "no"
