--- settings/installer/ospd/ssl/yes.yml
@@ -1,5 +1,0 @@
----
-
-installer:
-    overcloud:
-        use_ssl: "yes"
