--- settings/installer/ospd/storage/ceph.yml
@@ -1,12 +1,0 @@
----
-
-installer:
-    overcloud:
-        storage:
-            external: "no"
-            template:
-                content:
-                    parameter_defaults:
-                        CinderRbdPoolName: "volumes"
-                        NovaRbdPoolName: "vms"
-                        GlanceRbdPoolName: "images"
