--- settings/installer/ospd/storage/ceph/external/yes.yml
@@ -1,16 +1,0 @@
----
-
-installer:
-    overcloud:
-        storage:
-            external: "yes"
-            template:
-                content:
-                    parameter_defaults:
-                        CinderRbdPoolName: "automation-cinder"
-                        NovaRbdPoolName: "automation-nova"
-                        GlanceRbdPoolName: "automation-glance"
-                        CephClientUserName: "automation"
-                        CephClientKey: "{{ !lookup private.storage.ceph.CephClientKey }}"
-                        CephClusterFSID: "{{ !lookup private.storage.ceph.CephClusterFSID }}"
-                        CephExternalMonHost: "{{ !lookup private.storage.ceph.CephExternalMonHost }}"
