+++ settings/installer/ospd/undercloud/network/default.yml
@@ -1,0 +1,13 @@
+---
+undercloud_admin_vipundercloud_admin_vip: 172.16.0.11
+undercloud_public_vip: 172.16.0.10
+dhcp_start: 172.16.0.20
+dhcp_end: 172.16.0.120
+local_interface: eth0
+local_ip: 172.16.0.1/24
+masquerade_network: 172.16.0.0/24
+physical_network: datacentre
+network_gateway: 172.16.0.1
+discovery_iprange: 172.16.0.150,172.16.0.180
+network_cidr: 172.16.0.0/24
+network_type: flat
