--- settings/installer/ospd/version/7.yml
@@ -1,9 +1,0 @@
----
-
-installer:
-    version:
-        major: 7
-        minor: 0
-
-defaults:
-    minor: 3
