--- settings/installer/ospd/version/7/minor/1.yml
@@ -1,5 +1,0 @@
----
-
-installer:
-    version:
-        minor: 1
