--- settings/installer/ospd/version/7/minor/2.yml
@@ -1,5 +1,0 @@
----
-
-installer:
-    version:
-        minor: 2
