--- settings/installer/ospd/version/7/minor/3.yml
@@ -1,5 +1,0 @@
----
-
-installer:
-    version:
-        minor: 3
