--- settings/installer/ospd/version/8.yml
@@ -1,8 +1,0 @@
----
-
-installer:
-    version:
-        major: 8
-
-defaults:
-    minor: 0
