--- settings/installer/ospd/version/8/minor/0.yml
@@ -1,5 +1,0 @@
----
-
-installer:
-    version:
-        minor: 0
